他人のブログは削除できない

+++ app/Http/Controllers/Mypage/PostManageController.php
@@ -51,7 +51,6 @@ class PostManageController extends Controller
 
     public function update(Post $post, Request $request)
     {
-        // 所有チェック
         if (auth()->user()->isNot($post->user)) {
             abort(403);
         }
@@ -72,7 +71,9 @@ class PostManageController extends Controller
 
     public function destroy(Post $post)
     {
-        // 所有チェック
+        if (auth()->user()->isNot($post->user)) {
+            abort(403);
+        }
 
         $post->delete(); // 付随するコメントはDBの制約を使って削除する
